feat: 삭제 mutation 플로팅 버튼 안에서 호출되도록 수정
